:recycle: refact: alterando tipo do id do Chat para Long

+++ src/main/java/com/codifica/chatbot/core/application/chat_steps/cliente_event/WaitingForClienteEventStepHandler.java
@@ -45,7 +45,7 @@ public class WaitingForClienteEventStepHandler implements ConversationStep {
         }
     }
 
-    private void publishClienteEvent(Integer chatId, Map<String, String> dados) {
+    private void publishClienteEvent(Long chatId, Map<String, String> dados) {
         ClienteParaCadastrarEvent event = new ClienteParaCadastrarEvent(
                 chatId,
                 dados.get("nome"),
