+++ src/main/java/com/codifica/chatbot/core/application/chat_steps/pet_event/WaitingForPetEventStepHandler.java
@@ -69,7 +69,7 @@ public class WaitingForPetEventStepHandler implements ConversationStep {
         }
     }
 
-    private void publishPetEvent(Integer chatId, Integer clienteId, String petNome, Raca raca) {
+    private void publishPetEvent(Long chatId, Integer clienteId, String petNome, Raca raca) {
         PetParaCadastrarEvent event = new PetParaCadastrarEvent(chatId, clienteId, petNome, raca);
         petEventPublisherPort.publishPetParaCadastrar(event);
     }
